[AMQCPP-56] - fixing compiler warnings/errors on gcc
git-svn-id: https://svn.apache.org/repos/asf/incubator/activemq/activemq-cpp/trunk/activemq-cpp@500371 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/activemq/io/BufferedOutputStream.cpp
@@ -89,7 +89,7 @@ void BufferedOutputStream::flush() throw ( IOException ){
 ////////////////////////////////////////////////////////////////////////////////
 void BufferedOutputStream::write( const unsigned char c ) throw ( IOException ){
     
-    if( tail >= (int)bufferSize ){
+    if( tail >= bufferSize ){
         emptyBuffer();
     }
     
@@ -103,7 +103,7 @@ void BufferedOutputStream::write( const unsigned char* buffer, std::size_t len )
     // Iterate until all the data is written.
     for( std::size_t pos=0; pos < len; ){
         
-        if( tail >= (int)bufferSize ){
+        if( tail >= bufferSize ){
             emptyBuffer();
         }
